feat: :construction: Mejora de Arquitectura y Reducción de Redundancia de Código
Se finaliza la integracion de las vistas en visualizacion y edicion, falta terminar de cambiar los estilos a tailwind, algunos responsivos dañados.
