+++ src/components/camperProfile/styles/Proyects.module.css
@@ -13,6 +13,23 @@
     margin-top: 2rem;
   }
   
+  .projectCard {
+    width: 100%;
+    min-height: 450px;
+    height: 100%;
+    border-radius: 0.75rem;
+    overflow: hidden;
+  }
+
+  .projectCardAdd {
+    display: flex;
+    min-height: 450px;
+    width: 100%;
+    height: 100%;
+    align-items: center;
+    justify-content: center;
+  }
+
   .profileSubtitle {
     font-family: 'Poppins', sans-serif;
     font-weight: bold;
