--- src/components/camperProfileEdit/styles/ProyectsEdit.module.css
@@ -1,61 +1,0 @@
-.tecInfo {
-  position: relative;
-  width: 100%;
-}
-
-.projects {
-  position: relative;
-  z-index: 1;
-  display: grid;
-  grid-template-columns: repeat(4, 1fr);
-  gap: clamp(20px, 2vw, 20px);
-  width: 100%;
-  margin-top: 2rem;
-}
-
-.projectCard {
-  width: 100%;
-  min-height: 450px;
-  height: 100%;
-  border-radius: 0.75rem;
-  overflow: hidden;
-}
-
-.projectCardAdd {
-  display: flex;
-  min-height: 450px;
-  width: 100%;
-  height: 100%;
-  align-items: center;
-  justify-content: center;
-}
-
-.profileSubtitle {
-  font-family: "Poppins", sans-serif;
-  font-weight: bold;
-  color: #ffffff;
-  font-size: clamp(1.5rem, 3vw, 3rem);
-  margin-bottom: 1rem;
-  margin-top: 1rem;
-  display: block;
-  width: 100%;
-  position: relative;
-  z-index: 2;
-}
-
-.highlight {
-  color: #facc15;
-  font-weight: bold;
-}
-
-@media (max-width: 768px) {
-  .projects {
-    grid-template-columns: repeat(auto-fit);
-  }
-}
-
-@media (max-width: 480px) {
-  .projects {
-    grid-template-columns: 1fr;
-  }
-}
